Adds mpesa:install functionality to safely install the package
